backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
deepface
numpy
opencv-python
gunicorn
python-dotenv